Refactor code for OOP
