+++ src/main/java/Command.java
@@ -1,0 +1,3 @@
+public enum Command {
+        LIST, MARK, UNMARK, BYE, ADD, DELETE, UNKNOWN;
+}
