+++ src/main/java/MaxException.java
@@ -1,0 +1,5 @@
+public class MaxException extends Exception {
+    public MaxException(String msg) {
+        super(msg);
+    }
+}
